requirements.txt
numpy==2.2.4
pandas==2.2.3
openai==2.21.0
anthropic==0.89.0
google-genai==1.64.0